\name{acnr-package}
\alias{acnr-package}
\alias{acnr}
\docType{package}
\title{Annotated Copy-Number Regions}
\description{This data package contains SNP array data from different types of copy-number regions.  These regions were identified manually by the authors of the package and may be used to generate realistic data sets with known truth.}
\details{
\tabular{ll}{Package: \tab acnr\cr
Type: \tab Package\cr
Title: \tab Annotated Copy-Number Regions\cr
Version: \tab 0.2.2\cr
Date: \tab 2014-09-08\cr
Author: \tab Morgane Pierre-Jean and Pierre Neuvial\cr
Maintainer: \tab Morgane Pierre-Jean <morgane.pierrejean@genopole.cnrs.fr>\cr
License: \tab LGPL (>= 2.1)\cr
Depends: \tab R (>= 2.10), R.utils\cr
Suggests: \tab RUnit, BiocGenerics\cr
biocViews: \tab ExperimentData\cr}
}
\author{Morgane Pierre-Jean and Pierre Neuvial}

\keyword{ package }


