% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listDataSets.R
\name{listDataSets}
\alias{listDataSets}
\title{List available data sets}
\usage{
listDataSets()
}
\value{
name of one of the data sets of the package, see
  \link{listDataSets}
}
\description{
List available data sets
}
\examples{
listDataSets()

}

