% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listTumorFractions.R
\name{listTumorFractions}
\alias{listTumorFractions}
\title{List of available tumor fractions for a data set}
\usage{
listTumorFractions(dataSet)
}
\arguments{
\item{dataSet}{The name of a data set from the package, see \link{listDataSets}}
}
\value{
A numeric vector, the available tumor fractions for a data set
}
\description{
List of available tumor fractions for a data set
}
\examples{
dataSets <- listDataSets()
fracs <- listTumorFractions(dataSets[1])

}
